third_party/WebKit/LayoutTests/http/tests/devtools/sources/outline-javascript-6.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/sources-test.js"></script>

<script>

async function test() {
  var test = SourcesTestRunner.testJavascriptOutline;

  TestRunner.markStep('testClassAsNumberLiteralProperty');
  await test('var foo = { 42: class { }};');

  TestRunner.markStep('testClassAsStringLiteralProperty');
  await test('var foo = { "foo": class { }};');

  TestRunner.markStep('testClassAsIdentifierProperty');
  await test('var foo = { foo: class { }};');

  TestRunner.completeTest();
}

</script>

</head>

<body onload="runTest()">
<p>Verify javascript outline</p>
</body>
</html>
