third_party/WebKit/LayoutTests/http/tests/devtools/sources/outline-javascript-6-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verify javascript outline


Running: testClassAsNumberLiteralProperty
Text:
    var foo = { 42: class { }};
Outline:

Running: testClassAsStringLiteralProperty
Text:
    var foo = { "foo": class { }};
Outline:

Running: testClassAsIdentifierProperty
Text:
    var foo = { foo: class { }};
Outline:
    class foo:0:12
